backend/requirements.txt
flask
python-dotenv
psycopg[binary]
PyMySQL
